fix: registerated firebase database
